refactor get fav to cardCatelog
