+++ app/CardCatelog/CardCatelogActions.js
@@ -7,3 +7,10 @@ export const addFavorite = (favMov) => ({
   type: 'ADD_FAVORITE',
   favMov
 });
+
+export const getFavorites = (favoriteMovies) => {
+  return {
+    type: 'GET_FAVORITES',
+    favoriteMovies
+  };
+};
